+++ simple_history/__init__.py
@@ -1,4 +1,3 @@
-from __future__ import unicode_literals
 from pkg_resources import get_distribution, DistributionNotFound
 
 try: