+++ simple_history/tests/custom_user/admin.py
@@ -1,5 +1,3 @@
-from __future__ import unicode_literals
-
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin