+++ simple_history/tests/tests/utils.py
@@ -7,14 +7,9 @@ request_middleware = "simple_history.middleware.HistoryRequestMiddleware"
 
 OTHER_DB_NAME = "other"
 
-if django.__version__ >= "2.0":
-    middleware_override_settings = {
-        "MIDDLEWARE": (settings.MIDDLEWARE + [request_middleware])
-    }
-else:
-    middleware_override_settings = {
-        "MIDDLEWARE_CLASSES": (settings.MIDDLEWARE_CLASSES + [request_middleware])
-    }
+middleware_override_settings = {
+    "MIDDLEWARE": (settings.MIDDLEWARE + [request_middleware])
+}
 
 
 class TestDbRouter(object):